Cleanup & fixes
